+++ src/lib/Decklist.ts
@@ -1,4 +1,4 @@
-import { CardStorage, type PseudoCard, Card } from "./";
+import { CardStorage, type PseudoCard, Card } from ".";
 
 /**
  * A Decklist is a complete collection of Cards.
@@ -27,7 +27,8 @@ export class Decklist extends Array<Card> {
 	constructor(pseudoCards: Array<PseudoCard>) {
 		super();
 
-		this.defaultStorage = new CardStorage("deck");
+		this.#defaultStorage = new CardStorage("deck");
+
 		this.add(...pseudoCards);
 	}