+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 1.4.0
+
+- used webpack to add esm & cjs support
+- used webpack to add type declarations
+- moved assets
+
 ## 1.3.2
 
 - added decklist standard52_sextet
@@ -14,13 +20,13 @@
 - added events Card.onMoveStart and Card.onMoveEnd
 - changed build architecture
 
-## 1.0.2
+## 1.2.0
 
 - added standard52 images
 - fixed CardStorage.drawFrom positionning the receiving Card second-to-last instead of last
 - added a default export
 
-## 1.0.1
+## 1.1.0
 
 - added typing for intellisense